change many things: 1, fabs is changed to abs if argument is complex number. 2, remove MOTOR.cpp, induction_moter.cpp, and synchronous_motor.cpp of load models. 3, add define of and, or, not, in steps_enum.h for compling with Visual Studio. 4, large integer number such as 1e5 is changed to 100000 to eliminate warning of changing double to int. 5, some voltage variables were defined as size_t in bpa_powerflow_imexporter.cpp. change variable type to double. 6, add include<algorithm> when calling max and min for compliing with Visual Studio. 7, rename source.cpp in cpptest/ to source_cpptest.cpp. 8, std::advance failed in OWNERSHIP when trying to get owner and fraction with index greater than owner size. add comparision of index and owner size to fix the bug. 9, There was a 'winding' variable not initialized in psse_powerflow_imexporter.cpp when trying to export transformer data. Initialze it as PRIMARY_SIDE. 10, When solving powerflow to decoupled method, BQ cannot to properly LU factorized. The problem is caused by improper BQ matrix in which some entries are NAN. The NAN is caused when some FIXED_SHUNT has 0+j0 shunt power. If the power is 0, then the shunt impedance will be Infinite which is actually NAN in code, and the shunt admittance will be NAN too. Fix the bug by comparing the shunt power and handle the special case manually. 11. add dll exporation decalration for compiling with Visual Studio.

+++ code/source/basic/utility.cpp
@@ -34,7 +34,7 @@ string num2str(double number)
 
 int str2int(string str)
 {
-    return round(str2double(str));
+    return int(round(str2double(str)));
 }
 
 double str2double(string str)
